# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Read Station API",
        "description": "API for Read Station",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/bookcases": {
            "get": {
                "tags": [
                    "Admin / Bookcase"
                ],
                "summary": "Get all bookcases",
                "description": "Get all bookcases with pagination, search and filter",
                "operationId": "getAllBookcases",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Số trang hiện tại",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Số lượng mục trên mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Loại trạng thái",
                        "required": false,
                        "schema": {
                            "enum": [
                                "active",
                                "inactive",
                                "deleted"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all bookcases successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bookcases/{id}": {
            "get": {
                "tags": [
                    "Admin / Bookcase"
                ],
                "summary": "Get bookcase by id",
                "description": "Get bookcase by id",
                "operationId": "getBookcase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id của bookcase",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get bookcase successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Bookcase not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bookcases/create": {
            "post": {
                "tags": [
                    "Admin / Bookcase"
                ],
                "summary": "Create new bookcase",
                "description": "Create new bookcase",
                "operationId": "createBookcase",
                "requestBody": {
                    "description": "Create new bookcase",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bookcase_code",
                                    "description",
                                    "name"
                                ],
                                "properties": {
                                    "bookcase_code": {
                                        "type": "string",
                                        "default": null,
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create bookcase successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bookcases/update/{id}": {
            "put": {
                "tags": [
                    "Admin / Bookcase"
                ],
                "summary": "Update bookcase by id",
                "description": "Update bookcase by id",
                "operationId": "updateBookcase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id của bookcase",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update bookcase",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "name"
                                ],
                                "properties": {
                                    "bookcase_code": {
                                        "type": "string",
                                        "default": null,
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "default": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update bookcase successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Bookcase not found"
                    },
                    "500": {
                        "description": "Update bookcase failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bookcases/delete/{id}": {
            "delete": {
                "tags": [
                    "Admin / Bookcase"
                ],
                "summary": "Delete bookcase by id",
                "description": "Delete bookcase by id",
                "operationId": "deleteBookcase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id của bookcase",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete bookcase successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Bookcase not found"
                    },
                    "500": {
                        "description": "Delete bookcase failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}